+++ src/test/scala/io/viash/auxiliary/MainBuildAuxiliaryDockerChown.scala
@@ -19,7 +19,7 @@ class MainBuildAuxiliaryDockerChown extends FunSuite with BeforeAndAfterAll {
 
   def dockerChownGetOwner(dockerId: String): String = {
     val localConfig = configDockerOptionsChownFile
-    val localFunctionality = Config.read(localConfig, applyPlatform = false).functionality
+    val localFunctionality = Config.read(localConfig).functionality
     val localExecutable = Paths.get(tempFolStr, localFunctionality.name).toFile
 
     // prepare the environment
@@ -56,7 +56,7 @@ class MainBuildAuxiliaryDockerChown extends FunSuite with BeforeAndAfterAll {
 
   def dockerChownGetOwnerTwoOutputs(dockerId: String): (String,String) = {
     val localConfig = configDockerOptionsChownTwoOutputFile
-    val localFunctionality = Config.read(localConfig, applyPlatform = false).functionality
+    val localFunctionality = Config.read(localConfig).functionality
     val localExecutable = Paths.get(tempFolStr, localFunctionality.name).toFile
 
     // prepare the environment
@@ -97,7 +97,7 @@ class MainBuildAuxiliaryDockerChown extends FunSuite with BeforeAndAfterAll {
 
   def dockerChownGetOwnerMultipleOutputs(dockerId: String): (String,String,String) = {
     val localConfig = configDockerOptionsChownMultipleOutputFile
-    val localFunctionality = Config.read(localConfig, applyPlatform = false).functionality
+    val localFunctionality = Config.read(localConfig).functionality
     val localExecutable = Paths.get(tempFolStr, localFunctionality.name).toFile
 
     // prepare the environment